Backend/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
PyJWT
bcrypt
pydantic
pymongo
yfinance